Db details branch

+++ ui/db.py
@@ -1,6 +1,6 @@
 import sqlite3
 import hashlib
-import os
+
 
 
 class Database_Manager:
@@ -67,16 +67,18 @@ class Database_Manager:
     # CREATE(user_registration, add_book):
     def user_registration(self, username:str, password:str) -> bool:
         connection = self.create_connection()
+        hashed_password = hashlib.sha256(password.encode()).hexdigest()
         try:
              with connection:
                 cursor = connection.cursor()
                 new_user_insert = "INSERT INTO USERS (username, password) VALUES (?, ?)"
-                cursor.execute(new_user_insert,(username, password))#, (username, hashed_password))
-                connection.commit()
+                cursor.execute(new_user_insert,(username, hashed_password))
                 return True
+             
         except Exception as fail:
             print(f"ΣΦΑΛΜΑ ΚΑΤΑ ΤΗΝ ΕΓΓΡΑΦΗ ΤΟΥ ΧΡΗΣΤΗ: {fail}")
             return False
+        
         finally:
             connection.close()  
 
@@ -84,19 +86,22 @@ class Database_Manager:
 
         connection = self.create_connection()
         
-        title = book_dict.get('title', None)
-        authors = book_dict.get('authors', None)
-        year = book_dict.get('year', None)
+        title = book_dict.get('title')
+        authors = book_dict.get('authors')
+
+        if not title or not authors:
+            return None
+        
+        year = book_dict.get('year', 'Άγνωστη Ημερομηνία')
         isbn = book_dict.get('isbn', None)
-        description = book_dict.get('description', None)
+        description = book_dict.get('description', '')
         cover_img = book_dict.get('cover_img', None)
         volume_id = book_dict.get('volume_id', None)
-
 
         with connection:
             cursor = connection.cursor()
             try:
-                cursor.execute("SELECT book_id FROM BOOKS WHERE title = ? and authors = ?",(title, authors))
+                cursor.execute("SELECT book_id FROM BOOKS WHERE title = ? and authors = ?", (title, authors))
                 book_exists = cursor.fetchone()
                 if book_exists:
                     print("ΤΟ ΒΙΒΛΙΟ ΕΧΕΙ ΗΔΗ ΚΑΤΑΧΩΡΗΘΕΙ ΣΤΗΝ ΒΑΣΗ")
@@ -104,9 +109,8 @@ class Database_Manager:
                 
                 insert_query = "INSERT INTO BOOKS(title, authors, year, isbn, description, cover_img, volume_id) VALUES(?, ?, ?, ?, ?, ?, ?)"
                 cursor.execute(insert_query,(title, authors,year, isbn, description, cover_img, volume_id))
-
-                #connection.commit()
                 return cursor.lastrowid # Το ID του βιβλίου 
+            
             except Exception as fail:
                 print(f"ΑΠΟΤΥΧΙΑ ΕΙΣΑΓΩΓΗΣ: {fail}")
                 return None
@@ -116,19 +120,19 @@ class Database_Manager:
     # READ(find_user, get_book, search_books, get_all_books_with_stats, get_ratings):
     def find_user(self,username:str, password:str)-> int | None:
         connection = self.create_connection()
+        hashed_password = hashlib.sha256(password.encode()).hexdigest()
         try:             
             with connection:
                 cursor = connection.cursor()
 
                 search_query = "SELECT user_id from USERS WHERE username = ? AND password = ?"
-                cursor.execute(search_query,(username, password))
+                cursor.execute(search_query,(username, hashed_password))
 
                 user = cursor.fetchone()
-                #ΕΠΙΣΤΡΕΦΕΙ ΤΟ user_id Ή None
-            if user:
-                return user[0] 
-            else:
-                return None
+                return user[0] if user else None
+        except Exception as fail:
+            print(f"ΣΦΑΛΜΑ: {fail}")
+            return None
         finally:
             connection.close()
 
@@ -203,7 +207,7 @@ class Database_Manager:
                         cursor.execute(search_sql, (search_term, search_term, search_term))
 
                     rows = cursor.fetchall()
-                    #Επιστρέφουμε το αποτέλεσμα σε κανονικό λεξικό της Python
+                    # Επιστρέφουμε το αποτέλεσμα σε κανονικό λεξικό της Python
                     return [dict(row) for row in rows]
             
                 except Exception as fail:
@@ -213,34 +217,32 @@ class Database_Manager:
             connection.close()
 
     def get_all_books_with_stats(self)-> list[dict]:
-        '''Η ΣΥΝΑΡΤΗΣΗ ΜΑΣ ΕΠΙΣΤΡΕΦΕΙ ΤΟ ΣΥΝΟΛΟ ΤΩΝ ΒΙΒΛΙΩΝ ΜΕ ΤΟΝ ΜΕΣΟ ΟΡΟ ΤΩΝ ΑΞΙΟΛΟΓΗΣΕΩΝ
-        ΚΑΙ ΤΟΝ ΑΡΙΘΜΟ ΑΥΤΩΝ.
-        '''
         connection = self.create_connection()
         try:
             with connection:
                 connection.row_factory = sqlite3.Row
                 cursor = connection.cursor()
-                try:
-                    books_stats_query = '''
-                    SELECT BOOKS.*, 
-                    ROUND(AVG(RATINGS.rating), 2) AS avg_rate,
-                    COUNT(RATINGS.rating) AS total_rates
-                    FROM BOOKS
-                    LEFT JOIN RATINGS ON BOOKS.book_id = RATINGS.book_id
-                    GROUP BY BOOKS.book_id'''
                 
-                    cursor.execute(books_stats_query)
-                    rows = cursor.fetchall()
-                    all_book_stats = [] # ΑΡΧΙΚΟΠΟΙΗΣ ΛΙΣΤΑΣ
+                # ΧΡΗΣΗ LEFT JOIN ΚΡΑΤΑΜΕ ΟΛΑ ΤΑ ΔΕΔΟΜΕΝΑ ΑΚΟΜΑ ΚΑΙ ΓΙΑ ΒΙΒΛΙΑ ΧΩΡΙΣ ΒΑΘΜΟΛΟΓΙΑ
+                books_stats_query = '''
+                SELECT BOOKS.*, 
+                ROUND(AVG(RATINGS.rating), 2) AS avg_rate,
+                COUNT(RATINGS.rating) AS total_rates
+                FROM BOOKS
+                LEFT JOIN RATINGS ON BOOKS.book_id = RATINGS.book_id
+                GROUP BY BOOKS.book_id'''
+                
+                cursor.execute(books_stats_query)
+                rows = cursor.fetchall()
+                all_book_stats = [] # ΑΡΧΙΚΟΠΟΙΗΣ ΛΙΣΤΑΣ ΜΕ ΤΑ ΛΕΞΙΚΑ - ΕΓΓΡΑΦΕΣ
 
-                    for row in rows:
-                        all_book_stats.append(dict(row))
-                    return all_book_stats
+                for row in rows:
+                    all_book_stats.append(dict(row))
+                return all_book_stats
                 
-                except Exception as fail:
-                    print(f"ΣΦΑΛΜΑ ΕΥΡΕΣΗΣ ΣΤΑΤΙΣΤΙΚΩΝ: {fail}")
-                    return []
+        except Exception as fail:
+            print(f"ΣΦΑΛΜΑ ΕΥΡΕΣΗΣ ΣΤΑΤΙΣΤΙΚΩΝ: {fail}")
+            return []
         finally:
             connection.close()  
 
@@ -255,17 +257,18 @@ class Database_Manager:
 
                 read_books_query = '''SELECT BOOKS.*,
                                 RATINGS.rating,
-                                RATRINGS.comments
+                                RATINGS.comments
                                 FROM BOOKS
-                                JOIN RATINGS ON BOOKS.book_id = RATINGS.book_id
+                                INNER JOIN RATINGS ON BOOKS.book_id = RATINGS.book_id
                                 WHERE RATINGS.user_id = ?'''
 
-                cursor.execute(read_books_query, (user_id))
+                cursor.execute(read_books_query, (user_id,))
                 result = cursor.fetchall()
                 return [dict(r) for r in result]
             
         except Exception as fail:
                 print(f"Σφάλμα: {fail}")
+                return []
         finally:
                 connection.close()
     
@@ -320,20 +323,19 @@ class Database_Manager:
                     connection.commit()
                     return True
                 except sqlite3.IntegrityError as Error:
-                    print(f"ΠΡΟΕΚΥΨΕ ΑΚΕΡΑΙΟΤΗΤΑΣ: {Error}")
+                    print(f"ΣΦΑΛΜΑ ΑΚΕΡΑΙΟΤΗΤΑΣ: {Error}")
                 return False
         finally:
             connection.close()
 
     # DELETE:
     def delete_book(self, book_id: int)-> bool:
-        #Διαγραφή βιλίων και των αξιολογήσεων τους.
         connection = self.create_connection()
         
         try:
             with connection: 
                 cursor = connection.cursor()
-                #  Αναζήτηση του βιβλίου για να επιβεβαιώσουμε ότι υπάρχει πριν προσπαθήσουμε να το διαγράψουμε.
+                # Πρώτα select για να επιβεβαιώσουμε ότι υπάρχει το βιβλίο που θέλουμε να διαγράψουμε.
                 cursor.execute("SELECT * FROM BOOKS WHERE book_id = ?", (book_id,))
                 book_to_delete = cursor.fetchone()
                 if not book_to_delete:
@@ -344,7 +346,7 @@ class Database_Manager:
                 # Διαγραφή του βιβλίου από τον πίνακα BOOKS
                 cursor.execute("DELETE FROM BOOKS WHERE book_id = ?", (book_id,))
                 
-                return False
+                return True
                     
         except Exception as fail:
             print(f"ΣΦΑΛΜΑ ΚΑΤΑ ΤΗ ΔΙΑΓΡΑΦΗ: {fail}")
@@ -358,21 +360,20 @@ class Database_Manager:
         connection = self.create_connection()
         
         try:
-            with connection: # Transaction starts
+            with connection: 
                 cursor = connection.cursor()
                 # Αναζήτηση του χρήστη για να επιβεβαιώσουμε ότι υπάρχει πριν προσπαθήσουμε να τον διαγράψουμε.
                 cursor.execute("SELECT * FROM USERS WHERE user_id = ?", (user_id,))
                 user_to_delete = cursor.fetchone()
                 if not user_to_delete:
-                    print(f"ΠΛΗΡΟΦΟΡΙΑ: Δεν βρέθηκε χρήστης με ID {user_id}.")
+                    print(f"Δεν βρέθηκε χρήστης με ID {user_id}.")
                     return False    
-                # Διαγραφή των αξιολογήσεων που σχετίζονται με το χρήστη για να διατηρηθεί η ακεραιότητα των δεδομένων (Foreign Keys).
+                # Πρώτα διαγραφή των αξιολογήσεων που σχετίζονται με τον χρήστη για να διατηρηθεί η ακεραιότητα των δεδομένων (Foreign Keys).
                 cursor.execute("DELETE FROM RATINGS WHERE user_id = ?", (user_id,))
-                # Διαγραφή του χρήστη από τον πίνακα USERS
+                # Έπειτα διαγραφή του χρήστη από τον πίνακα USERS
                 cursor.execute("DELETE FROM USERS WHERE user_id = ?", (user_id,))
                 
-                return True
-                    
+                return True      
         except Exception as fail:
             print(f"ΣΦΑΛΜΑ ΚΑΤΑ ΤΗ ΔΙΑΓΡΑΦΗ ΧΡΗΣΤΗ: {fail}")
             return False
@@ -387,19 +388,17 @@ class Database_Manager:
         try:
             with connection: 
                 cursor = connection.cursor()
-                #Ευρεση της αξιολόγησης για να επιβεβαιώσουμε ότι υπάρχει πριν προσπαθήσουμε να την διαγράψουμε.
+                # Πρώτα select για να επιβεβαιώσουμε ότι υπάρχει η αξιολόγηση που θέλουμε να διαγράψουμε.
                 cursor.execute("SELECT * FROM RATINGS WHERE user_id = ? AND book_id = ?", (user_id, book_id))
                 rating_to_delete = cursor.fetchone()
 
                 if not rating_to_delete:
-                    print(f"ΠΛΗΡΟΦΟΡΙΑ: Δεν βρέθηκε αξιολόγηση με ID χρήστη {user_id} και ID βιβλίου {book_id}.")
+                    print(f"Δεν βρέθηκε αξιολόγηση με ID χρήστη {user_id} και ID βιβλίου {book_id}.")
                     return False
                 
-                cursor.execute(
-                    "DELETE FROM RATINGS WHERE user_id = ? AND book_id = ?", 
-                    (user_id, book_id))
-                return True
-                    
+                cursor.execute("DELETE FROM RATINGS WHERE user_id = ? AND book_id = ?", (user_id, book_id))
+                return True  
+            
         except Exception as fail:
             print(f"ΣΦΑΛΜΑ ΚΑΤΑ ΤΗ ΔΙΑΓΡΑΦΗ ΑΞΙΟΛΟΓΗΣΗΣ: {fail}")
             return False
